refactor(sense): remove ProcessBuilder top header per redesign (sidebar+content+footer only)
